Fix TUI prompt input render performance
